requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
docopt==0.6.2
idna==3.4
mysqlclient==2.1.1
Pillow==9.3.0
pipreqs==0.4.11
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
yarg==0.1.9
npm==0.1.1
graphene_django==3.0.0
async-graphql==0.1.1
graphene-file-upload==1.3.0
django-multiselectfield==0.1.12